feat: migrate old methods (without much refactoring)
